Cool Down adjustment
Adjusted the cooldown. When there was still data being transmitted at 0 watts it would interrupt the cooldown adjuster and would keep calculating and adjustment of 0. It was changed to be 1/10th with rounding up to ensure it goes down. The exactness of this math is poor, but cooldown isn't an exact science for timing.

Additional bug with triac_set_normal was set to false, and there was no powermeter connected it would default to the last offset setting since there was no calculation or re-enter to set_power.  Calling set_power with 0 caused some sort of loop instability that was unpredictable, so offset was manually set to 500